requirements.txt
asgiref==3.2.3
autopep8==1.4.4
coverage==5.0.1
defusedxml==0.4.1
dj-database-url==0.5.0
Django==3.0.1
django-crispy-forms==1.8.1
django-environ==0.4.5
django-filter==2.2.0
django-heroku==0.3.1
django-markdownx==3.0.0
django-test==0.4030
gunicorn==20.0.4
lxml==4.4.2
Markdown==3.1.1
Pillow==6.2.1
psycopg2==2.8.4
pycodestyle==2.5.0
pytz==2019.3
selenium==3.141.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==5.0.1
